loader: comparison between signed and unsigned integer expressions
/code/illumos-gate/usr/src/boot/lib/libstand/rpc.c: In function 'rpc_call':
/code/illumos-gate/usr/src/boot/lib/libstand/rpc.c:178:9: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (cc <= sizeof(*reply)) {
         ^~
/code/illumos-gate/usr/src/boot/lib/libstand/rpc.c: In function 'rpc_getport':
/code/illumos-gate/usr/src/boot/lib/libstand/rpc.c:415:9: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  if (cc < sizeof(*res)) {
         ^
cc1: all warnings being treated as errors

+++ usr/src/boot/lib/libstand/rpc.c
@@ -175,7 +175,7 @@ rpc_call(struct iodesc *d, n_long prog, n_long vers, n_long proc,
 	if (cc == -1)
 		return (-1);
 
-	if (cc <= sizeof(*reply)) {
+	if (cc <= (ssize_t)sizeof(*reply)) {
 		errno = EBADRPC;
 		free(ptr);
 		return (-1);
@@ -413,7 +413,7 @@ rpc_getport(struct iodesc *d, n_long prog, n_long vers)
 
 	cc = rpc_call(d, PMAPPROG, PMAPVERS, PMAPPROC_GETPORT,
 	    args, sizeof(*args), (void **)&res, &pkt);
-	if (cc < sizeof(*res)) {
+	if (cc < (ssize_t)sizeof(*res)) {
 		printf("getport: %s", strerror(errno));
 		errno = EBADRPC;
 		free(pkt);
